feat: rename Navbar to NavigationMenu and create Navbar.astro
