+++ src/components/Home/Hero.astro
@@ -8,7 +8,7 @@ import calendar from '@/assets/icons/calendar.svg';
 import convert from '@/assets/icons/convert.svg';
 ---
 <section id="hero" class="flex flex-col sm:gap-10">
-	<a href="/order">
+	<a href="/order" class="sm:w-[1140px] sm:h-[357px]">
 		<Image src={jumbotron} alt="Dapatkan nomor cantik 11 digit" class="hidden sm:block" />
 		<Image src={mobileBanner} alt="Dapatkan nomor cantik 11 digit" class="block sm:hidden" />
 	</a>
