+++ src/components/ui/DropdownMenu/DropdownMenu.types.ts
@@ -1,6 +1,7 @@
-import type { MenuItem } from "@/components/ui/Navbar/Navbar.types";
+import type { MenuItem } from "@/components/ui/NavigationMenu/NavigationMenu.types";
 
 export interface DropdownMenuProps {
   buttonLabel: string;
   options: MenuItem[];
+  isPopular?: boolean;
 }
