--- src/components/ui/Navbar/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Navbar";
