+++ src/components/ui/NavigationMenu/NavigationMenu.types.ts
@@ -2,4 +2,5 @@ export interface MenuItem {
   label: string;
   href?: string;
   subMenu?: MenuItem[];
+  isPopular?: boolean;
 }
