+++ src/layouts/Footer.astro
@@ -49,7 +49,7 @@ import instagram from '@/assets/icons/instagram.svg';
 			</a>
 			<a
 				href="https://tiktok.com/IndosatOoredooHutchison" 
-				class="hidden sm:block h-8 w-8 rounded-full flex justify-center items-center"
+				class="hidden sm:flex h-8 w-8 rounded-full justify-center items-center"
 				style={{ backgroundColor: "#E8E8E8" }}
 			>
 				<Image src={tiktok} alt="Tiktok icon" />
