+++ src/layouts/Layout.astro
@@ -1,6 +1,6 @@
 ---
-import Navbar from '@/components/ui/Navbar';
-import Footer from '@/components/ui/Footer.astro'
+import Navbar from '@/layouts/Navbar.astro';
+import Footer from '@/layouts/Footer.astro'
 interface Props {
 	title: string;
 }
@@ -19,7 +19,7 @@ const { title } = Astro.props;
 		<title>{title}</title>
 	</head>
 	<body class="bg-ioh-neutral-50">
-		<Navbar client:load/>
+		<Navbar />
 		<main class="container sm:max-w-[1140px] mx-auto flex flex-col gap-6 sm:gap-16">
 			<slot />
 		</main>
